Simple docker image and instructions

+++ src/main/java/com/noodlesscoders/forumxbackend/api/message/bean/MessageOB.java
@@ -4,27 +4,27 @@ import java.util.Date;
 
 public class MessageOB {
 
-    private Integer id;
+    private Long id;
 
-    private Integer userID;
+    private Long userID;
 
     private String message;
 
     private Date createTime;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getUserID() {
+    public Long getUserID() {
         return userID;
     }
 
-    public void setUserID(Integer userID) {
+    public void setUserID(Long userID) {
         this.userID = userID;
     }
